% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_function.R
\name{anova_rcbd}
\alias{anova_rcbd}
\title{Anova analysis for given AST204 problem
anova_rcbd gets the anova table for given dataset whose treatments are given
rowwise and blocks will be columnwise}
\usage{
anova_rcbd(data)
}
\arguments{
\item{data}{}
}
\value{
dataframe
}
\description{
Anova analysis for given AST204 problem
anova_rcbd gets the anova table for given dataset whose treatments are given
rowwise and blocks will be columnwise
}
\examples{
obs = c(25,35,21,37,40,52,39,58,27,34,20,31)
a = 3
b = 4
N = 12
mat = matrix(obs, nrow = a, ncol = b, byrow = T)
anova_rcbd(mat)
}
